add support for locating css, jpeg, ttf and xml resources from their resource files' contentType
This is obvisouly not exhaustive, and there is no requirement that files of a particular contentType must have a specific file extension, there is only convention, so mismataches may occur in the future.

A better long-term strategy may be to reverse the process, looking up the resource-meta.xml file from the resource file name, as this is a much simpler and consistent algorithm.